style: 코드 간결화, 컨벤션 적용, dto 디렉토리 구성
