+++ src/main/kotlin/com/devfloor/untitled/article/application/request/ArticleModifyRequest.kt
@@ -1,4 +1,4 @@
-package com.devfloor.untitled.article.application
+package com.devfloor.untitled.article.application.request
 
 data class ArticleModifyRequest(
     val options: List<String>,
