+++ src/main/kotlin/com/devfloor/untitled/article/application/response/ArticleFeedResponse.kt
@@ -1,24 +1,22 @@
-package com.devfloor.untitled.article.application
+package com.devfloor.untitled.article.application.response
 
 import com.devfloor.untitled.article.domain.Article
 import com.devfloor.untitled.articlefavorite.domain.ArticleFavorite
 import com.devfloor.untitled.articlefavorite.domain.ArticleFavoriteType
-import com.devfloor.untitled.articlehashtag.domain.ArticleHashtag
 import com.devfloor.untitled.articleoption.domain.ArticleOption
-import com.devfloor.untitled.user.application.ProfileResponse
 import com.fasterxml.jackson.annotation.JsonFormat
 import java.time.LocalDateTime
 
-data class ArticleResponse(
+data class ArticleFeedResponse(
+    val id: Long,
+
     val options: List<String>,
 
     val title: String? = null,
 
     val anonymous: Boolean,
 
-    val content: String,
-
-    val author: ProfileResponse,
+    val authorName: String,
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", shape = JsonFormat.Shape.STRING)
     val createdDate: LocalDateTime,
@@ -26,7 +24,7 @@ data class ArticleResponse(
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", shape = JsonFormat.Shape.STRING)
     val modifiedDate: LocalDateTime,
 
-    val hashtags: List<String>,
+    val content: String,
 
     val favorites: Long,
 
@@ -37,19 +35,22 @@ data class ArticleResponse(
     constructor(
         articleOptions: List<ArticleOption>,
         article: Article,
-        articleHashtags: List<ArticleHashtag>,
         articleFavorites: List<ArticleFavorite>,
     ) : this(
-        options = articleOptions.map { it.option.type.name },
+        id = article.id,
+        options = articleOptions.map { it.option.type.text },
         title = article.title,
         anonymous = article.anonymous,
-        content = article.content,
-        author = ProfileResponse(article.author),
+        authorName = article.author.nickname,
         createdDate = article.createdDate,
         modifiedDate = article.modifiedDate,
-        hashtags = articleHashtags.map { it.hashtag.name },
+        content = article.sliceContentByLength(CONTENT_MAX_LENGTH),
         favorites = articleFavorites.count { it.matchType(ArticleFavoriteType.FAVORITE) }.toLong(),
         wonders = articleFavorites.count { it.matchType(ArticleFavoriteType.WONDER) }.toLong(),
         clips = articleFavorites.count { it.matchType(ArticleFavoriteType.CLIP) }.toLong(),
     )
+
+    companion object {
+        private const val CONTENT_MAX_LENGTH: Long = 300
+    }
 }
