+++ src/main/kotlin/com/devfloor/untitled/article/application/response/ArticleResponse.kt
@@ -1,22 +1,24 @@
-package com.devfloor.untitled.article.application
+package com.devfloor.untitled.article.application.response
 
 import com.devfloor.untitled.article.domain.Article
 import com.devfloor.untitled.articlefavorite.domain.ArticleFavorite
 import com.devfloor.untitled.articlefavorite.domain.ArticleFavoriteType
+import com.devfloor.untitled.articlehashtag.domain.ArticleHashtag
 import com.devfloor.untitled.articleoption.domain.ArticleOption
+import com.devfloor.untitled.user.application.ProfileResponse
 import com.fasterxml.jackson.annotation.JsonFormat
 import java.time.LocalDateTime
 
-data class ArticleFeedResponse(
-    val id: Long,
-
+data class ArticleResponse(
     val options: List<String>,
 
     val title: String? = null,
 
     val anonymous: Boolean,
 
-    val authorName: String,
+    val content: String,
+
+    val author: ProfileResponse,
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", shape = JsonFormat.Shape.STRING)
     val createdDate: LocalDateTime,
@@ -24,7 +26,7 @@ data class ArticleFeedResponse(
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", shape = JsonFormat.Shape.STRING)
     val modifiedDate: LocalDateTime,
 
-    val content: String,
+    val hashtags: List<String>,
 
     val favorites: Long,
 
@@ -35,22 +37,19 @@ data class ArticleFeedResponse(
     constructor(
         articleOptions: List<ArticleOption>,
         article: Article,
+        articleHashtags: List<ArticleHashtag>,
         articleFavorites: List<ArticleFavorite>,
     ) : this(
-        id = article.id,
-        options = articleOptions.map { it.option.type.text },
+        options = articleOptions.map { it.option.type.name },
         title = article.title,
         anonymous = article.anonymous,
-        authorName = article.author.nickname,
+        content = article.content,
+        author = ProfileResponse(article.author),
         createdDate = article.createdDate,
         modifiedDate = article.modifiedDate,
-        content = article.sliceContentByLength(CONTENT_MAX_LENGTH),
+        hashtags = articleHashtags.map { it.hashtag.name },
         favorites = articleFavorites.count { it.matchType(ArticleFavoriteType.FAVORITE) }.toLong(),
         wonders = articleFavorites.count { it.matchType(ArticleFavoriteType.WONDER) }.toLong(),
         clips = articleFavorites.count { it.matchType(ArticleFavoriteType.CLIP) }.toLong(),
     )
-
-    companion object {
-        private const val CONTENT_MAX_LENGTH: Long = 300
-    }
 }
