+++ src/main/kotlin/com/devfloor/untitled/board/presentation/BoardController.kt
@@ -1,7 +1,7 @@
 package com.devfloor.untitled.board.presentation
 
 import com.devfloor.untitled.board.application.BoardFacade
-import com.devfloor.untitled.board.application.BoardResponse
+import com.devfloor.untitled.board.application.response.BoardResponse
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.ResponseStatus
